Slurm autoscaling: Splitting gridengine.py into several files by engine type
