+++ workflows/pipe-common/test/test_scale_down.py
@@ -18,7 +18,7 @@ from datetime import datetime, timedelta
 from mock import MagicMock, Mock
 
 from pipeline.hpc.autoscaler import GridEngineAutoscaler
-from pipeline.hpc.gridengine import GridEngineJob, GridEngineJobState
+from pipeline.hpc.engine.gridengine import GridEngineJob, GridEngineJobState
 from pipeline.hpc.host import MemoryHostStorage
 from pipeline.hpc.utils import Clock
